Improve relative path code in vendor updater

+++ common/spec/dependabot/file_updaters/vendor_updater_spec.rb
@@ -30,6 +30,8 @@ RSpec.describe Dependabot::FileUpdaters::VendorUpdater do
     before do
       in_cloned_repository(repo_contents_path) do
         # change a vendor file like an updater would
+        next unless File.exist?("vendor/cache/business-1.4.0.gem")
+
         `mv vendor/cache/business-1.4.0.gem vendor/cache/business-1.5.0.gem`
       end
     end
@@ -94,7 +96,8 @@ RSpec.describe Dependabot::FileUpdaters::VendorUpdater do
       before do
         in_cloned_repository(repo_contents_path) do
           # change a vendor file like an updater would
-          `mv nested/vendor/cache/business-1.4.0.gem nested/vendor/cache/business-1.5.0.gem`
+          `mv nested/vendor/cache/business-1.4.0.gem \
+          nested/vendor/cache/business-1.5.0.gem`
         end
       end
